Better support for ID-less entities, with new +shouldAlwaysCreateNew option
When merging data from the server, this option disables the checks to see if the data already exists locally, and instead assumes it doesn't exist, and always creates a new record.

This is useful when working with servers that do not provide a unique primary key for certain entities, and cannot be changed to do so.

It is still recommended to have a primary key, if possible.

See TestEntityFour as an example for how to use en entity without an ID.